requirements.txt
###### Requirements without Version Specifiers ######`

###### Requirements with Version Specifiers ######`
setuptools==69.0.3
cc_tools @ git+https://github.com/ChipMcCallahan/CCTools.git

pygame~=2.5.2
pillow~=10.2.0
